Move pkg/repositories to pkg/input/repositories
