+++ pkg/input/repositories/testdata/artificial.yaml
