+++ pkg/input/repositories/testdata/team-atlas.yaml
