+++ pkg/input/repositories/testdata/team-clippy.yaml
