+++ pkg/legacy/create-component-entity.go
@@ -9,7 +9,7 @@ import (
 	bscatalog "github.com/giantswarm/backstage-catalog-importer/pkg/bscatalog/v1alpha1"
 	"github.com/giantswarm/backstage-catalog-importer/pkg/catalog/component"
 	"github.com/giantswarm/backstage-catalog-importer/pkg/input/helmchart"
-	"github.com/giantswarm/backstage-catalog-importer/pkg/repositories"
+	"github.com/giantswarm/backstage-catalog-importer/pkg/input/repositories"
 )
 
 // CreateComponentEntity is the deprecated way of generating a component entity.
